fix(ui): align with monorepo + core package conventions
- package.json: drop dead `^1` in lucide-react peerDep range
- primitives.tsx: extract Button/INPUT_CLASS/LABEL_CLASS out of templates
- context.ts: drop [key: string]: unknown index signature, typed API
- index.tsx: memoize ctx so downstream useMemo actually memoize
- api.ts: export UPLOADS_PATH, consume at 3 SchemaForm call sites
- use-schema-form-labels: tighten return type to Required + addItem
