+++ packages/ui/src/schema-form/context.ts
@@ -13,5 +13,4 @@ export interface SchemaFormContext {
   uploadPath?: string;
   upload?: UploadFn;
   labels?: FileWidgetLabels & { addItem?: string };
-  [key: string]: unknown;
 }
